Add AVX2 [inv]NTT from official Kyber repository (#203)
* Add AVX2 NTT from official Kyber implementation

The AVX2 Kyber NTT produces non-reduced outputs in
non-standard order, for better performance. Neither
is currently supported by the C<->Native interface.

To accommodate, we temporarily reduce and reorder the
output of the AVX2 NTT to match the current C<->Native
interface. This will likely need changing later.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>

* Increase default alignment of buffers to 32 bytes

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>

* Add AVX2 invNTT from official Kyber implementation

As for the forward NTT, the inverse NTT operates on
a non-standard order that currently supported by our
C<->Native interface. To accommodate, we shuffle the
data in the expected order first before calling the
AVX2 iNTT. This may need revisiting later for better
performance.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>

---------

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>