CI: Use -O0 instead of -O3 for CI tests
O3 slows down CI runs, and is not needed except when we benchmark
for performance.
This commit explicitly sets -O0 for the compatibility and
compiler tests to speed-up CI.
This also adds an additional variable to pass to the Makefile
called EXTRAFLAGS which can be used to overwrite default flags.
Those will be appended to the CFLAGS after the default flags.
Otherwise, the default optimization flags overwrite the
custom flags.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>
Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>