Tests: Remove expect_proc vs. actual_proc distinction
The test script processes the output of test runs with two functions:
An `actual_proc` and an `expect_proc`. The `actual_proc` is mostly
preprocessing, while the `expect_proc` checks against an expected value.

The does not seem to be a reason for this split. This commit conatenates
`actual_proc` and `expect_proc` into a single output-checking function
`check_proc`.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>