Add KAT checks to `make quickcheck`
This commit adds a basic script META.sh to query the META.json.
If jq is present, it's used for the query. Otherwise, some basic
fiddling with cat, grep, cut, and tr is used.

This is used to conduct a bash-based KAT test in `make quickcheck`,
rather then running the test scripts, which already require a Python
setup.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>