Build: Fix use of CFLAGS when checking for duplicated symbols
When we build archive files for mlkem-native, we check for duplicated
symbols by linking the target library against a dummy .c file. The
ad-hoc compiler invocation used for this did not use and `-z noexecstack`
flag which led to linker warnings.

This commit addresses this by passing CFLAGS to the dummy compilation,
which in particular includes `-z noexecstack`. The only exception is
`-static`, which is part of the CFLAGS for the library build, but dropped
for the dummy compilation.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>