Remove ARCH_FLAGS from test script and Makefile
Previously, we distinguished between CFLAGS and ARCH_FLAGS in the
`Makefile` and the `tests` script. Ultimately, however, they are
both just flags passed to the compiler, and the semantic difference
is not worth highlighting at the level of the Makefile and tests
script. This commit thus removes ARCH_FLAGS and uses CFLAGS instead.

The CI workflows so far retain their `archflags` parameter, but it
is passed as `--cflags` to the `tests` script.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>