Tests: Run ACVP test via make
A previous commit has re-introduced the standalone ACVP client
acvp_client.py, but the `tests` script was not yet changed to use it.

This commit removes the duplication of ACVP checking logic between
`tests` and `acvp_client.py` by having `tests` call `make check_acvp`
instead.

For now, this means that the distinction of ACVP test type and scheme
is gone, and we only see a single line of test output for all ACVP tests.
This may be improved in the future by adding a corresponding parameter
to `acvp_client.py`.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>