acvp_client.py: Support cmd prefix using EXEC_WRAPPER env variable
When using cross-compilation & emulation, `acvp_client.py` fails
because it tries to run the ACVP client natively.

This commit provides the basis for fixing this by checking for
the environment variable `EXEC_WRAPPER` which, if present, should
be prefix the binary invocation. This will normally be set to some
QEMU binary.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>