Makefile: Fix race condition during duplicated symbol checking
When checking for duplicated symbols, we build a temporary `a.out`
binary by linking a dummy C file against libmlkem{512,768,1024}.a.
Afterwards, `a.out` is removed.

When running multiple `make` jobs in parallel, this can lead to
a race condition where multiple jobs create and destroy `a.out`
at the same time.

This commit fixes this by using a temporary filename derived from
the library target, hence avoiding the nameclash and race condition.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>